move ask() to nicebox.c
