+++ src/Makefile.am
@@ -16,7 +16,7 @@ tlf_SOURCES = \
 	fldigixmlrpc.c freq_display.c focm.c \
 	genqtclist.c \
 	get_time.c getctydata.c getexchange.c getmessages.c getpx.c \
-	getsummary.c gettxinfo.c getwwv.c grabspot.c \
+	gettxinfo.c getwwv.c grabspot.c \
 	initial_exchange.c \
 	keyer.c \
 	lancode.c last10.c listmessages.c log_to_disk.c log_utils.c \
@@ -53,7 +53,7 @@ noinst_HEADERS = \
 	fldigixmlrpc.h freq_display.h focm.h \
 	genqtclist.h \
 	get_time.h  getctydata.h getexchange.h getmessages.h getpx.h \
-	getsummary.h gettxinfo.h getwwv.h globalvars.h grabspot.h \
+	gettxinfo.h getwwv.h globalvars.h grabspot.h \
 	ignore_unused.h initial_exchange.h \
 	keyer.h keystroke_names.h \
 	lancode.h last10.h listmessages.h log_utils.h \
